requirements.txt
Flask==2.3.3
requests==2.31.0
cryptography==41.0.7
assemblyai==0.23.1
python-dotenv==1.0.1
python-docx==0.8.12